[fix] separate scheduler test into different file, remove unwanted code comments
